trading-chatbot/requirements.txt
aiohttp==3.9.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
asyncio==3.4.3
attrs==25.4.0
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.15.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
boto3==1.41.2
botocore==1.41.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
curl_cffi==0.13.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.1
faiss-cpu==1.13.0
fastapi==0.121.3
fastjsonschema==2.21.2
filelock==3.20.0
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
fyers_apiv3==3.1.7
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lark==1.3.1
llvmlite==0.44.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook_shim==0.2.4
nsepy==0.8
numba==0.61.2
numpy==2.2.6
openai==2.8.1
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
pandocfilters==1.5.1
parso==0.8.5
peewee==3.18.3
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.3
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.29.0
s3transfer==0.15.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.6.1
websockets==15.0.1
wrapt==2.0.1
yarl==1.22.0
yfinance==0.2.66
